1.6.47
* Добавлен вывод информации по ресурсам процесса
* Доработан вывод импорта/экспорта ELF файлов в плагине dwarfreader
* Для части модулей (ELF) добавлена поддержка Lazarus для использования в Linux
* Загрузка отладочной информации теперь настраиваемая через TImageLoadType

+++ MemoryMap/MemoryMap.DebugMapData.pas
@@ -39,7 +39,7 @@ type
   private type
     TSectionData = record
       Index: Integer;
-      StartAddr: DWORD;
+      StartAddr: UInt64;
       Length: DWORD;
       Executable: Boolean;
     end;
